refactoring + created a class to contain the utility functions
